Refactor code for uploading videos and images
This commit refactors the code for uploading videos and images in the `PostReposatory` and `AddPostCubit` classes. The `PostReposatory` now has a new method `getUploadUrl` that retrieves the upload URL for a video file. The `AddPostCubit` class now uses this method to upload videos and images. The `UploadImageLoading` state has been renamed to `UploadVideoLoading` to reflect this change.