Refactor AddPostCubit to handle video posts separately
The AddPostCubit has been refactored to handle video posts separately from other types of posts. This was done by creating new state classes and updating the createVideoPost method to emit the appropriate state. Additionally, the clearpostContant method was updated to clear the postImages list when creating a video post. The CreatePostView was also updated to handle the new video post states.