Refactor code for better readability and maintainability
This commit refactors the code in several files for better readability and maintainability. In the `LastPostsContainer` widget, `SizedBox` widgets are added to wrap `CustomViewPostPhoto` widgets to ensure consistent spacing. In the `SettingsRepository`, the return type of the `mySubscriptePlan` method is changed to `List<MySubscriptionModel>` and a new method `myPlan` is added to retrieve the user's subscription plan. In the `SettingsCubit`, the `mySubscription` method is renamed to `myPlan` to reflect the new method name in the repository and the state classes are updated accordingly. Finally, in the `SettingsScreen`, the `mySubscription` method call is replaced with `myPlan` to match the updated method name.